requirements.txt
aiogram~=3.15.0
asyncpg~=0.30.0
python-decouple~=3.8